requirements.txt
astroid==2.3.3
colorama==0.4.3
Django==2.1.5
django-crispy-forms==1.8.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
psycopg2-binary==2.8.4
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
pdf2image==1.13.1
requests==2.23.0
gunicorn==20.0.4
nltk==3.5
scikit-learn==0.23.2
pyspellchecker==0.5.4